refactor(settings): move telemetry field
