+++ app/react/portainer/settings/SettingsView/ApplicationSettingsPanel/ApplicationSettingsPanel.tsx
@@ -7,11 +7,9 @@ import {
   useUpdateSettingsMutation,
 } from '@/react/portainer/settings/queries';
 import { notifySuccess } from '@/portainer/services/notifications';
-import { useIsDemo } from '@/react/portainer/system/useSystemStatus';
 
 import { Widget } from '@@/Widget';
 import { LoadingButton } from '@@/buttons';
-import { SwitchField } from '@@/form-components/SwitchField';
 import { FormControl } from '@@/form-components/FormControl';
 import { Input } from '@@/form-components/Input';
 
@@ -22,7 +20,7 @@ import { Values } from './types';
 import { LogoFieldset } from './LogoFieldset';
 import { ScreenBannerFieldset } from './ScreenBannerFieldset';
 import { TemplatesUrlSection } from './TemplatesUrlSection';
-import { DemoAlert } from './DemoAlert';
+import { EnableTelemetryField } from './EnableTelemetryField';
 
 export function ApplicationSettingsPanel({
   onSuccess,
@@ -90,7 +88,6 @@ export function ApplicationSettingsPanel({
 
 function InnerForm({ isLoading }: { isLoading: boolean }) {
   const { values, setFieldValue, isValid, errors } = useFormikContext<Values>();
-  const isDemoQuery = useIsDemo();
 
   return (
     <Form className="form-horizontal">
@@ -119,32 +116,7 @@ function InnerForm({ isLoading }: { isLoading: boolean }) {
 
       <LogoFieldset />
 
-      <div className="form-group">
-        <div className="col-sm-12">
-          <SwitchField
-            labelClass="col-sm-2"
-            label="Allow the collection of anonymous statistics"
-            checked={values.enableTelemetry}
-            name="toggle_enableTelemetry"
-            onChange={(checked) => setFieldValue('enableTelemetry', checked)}
-            disabled={isDemoQuery.data}
-          />
-        </div>
-
-        <DemoAlert />
-
-        <div className="col-sm-12 text-muted small mt-2">
-          You can find more information about this in our{' '}
-          <a
-            href="https://www.portainer.io/documentation/in-app-analytics-and-privacy-policy/"
-            target="_blank"
-            rel="noreferrer"
-          >
-            privacy policy
-          </a>
-          .
-        </div>
-      </div>
+      <EnableTelemetryField />
 
       <ScreenBannerFieldset />
